move application config and constance to main folder as they are not shared
