+++ src/main/application-config.js
